Update getRetainedSize methods to return long

+++ src/main/java/io/airlift/slice/BasicSliceOutput.java
@@ -62,7 +62,7 @@ public class BasicSliceOutput
     }
 
     @Override
-    public int getRetainedSize()
+    public long getRetainedSize()
     {
         return slice.getRetainedSize() + INSTANCE_SIZE;
     }
